Pause menu created, Temporary inventory created, new HP system implemented.
I have created a pause menu in game and an inventory system in which when brought up it will show your HP and the amount of each type of health packs that you have as well as a basic text element for the time. There are a couple of bugs that need to be adressed as well as further development of what I have currently implemented. TODO's ahve been stated in the relevent scripts.