Small changes regarding Impact Prefab
Increased the range of the weapons (not to infinite because that's slower to RayCast).

Utilised a Pool for the Impact_Prefabs because constantly instantiating and destroying objects has an inherent memory leak in Unity.

Changed how Health Packs work, but I realise that the idea is that they should go into the player's inventory so this will need to be fixed later on.